Move *Annotation classes under core/annotation
We usually group similar files in the same directories.
