+++ lightly_studio/src/lightly_studio/core/annotation/__init__.py
@@ -1,0 +1,3 @@
+from .annotation import Annotation
+
+__all__ = ["Annotation"]
