+++ lightly_studio/src/lightly_studio/core/annotation/annotation.py
