+++ lightly_studio/src/lightly_studio/core/annotation/instance_segmentation.py
