+++ lightly_studio/src/lightly_studio/core/annotation/object_detection.py
